TxEssence trait
