+++ primitives/src/lib.rs
@@ -18,6 +18,7 @@ pub mod access_list;
 pub mod block;
 pub mod keccak;
 pub mod receipt;
+pub mod signature;
 pub mod transactions;
 pub mod trie;
 pub mod withdrawal;
