+++ testing/ef-tests/src/lib.rs
@@ -37,8 +37,9 @@ use zeth_primitives::{
     ethers::from_ethers_h160,
     keccak::keccak,
     revm::from_revm_b160,
+    signature::TxSignature,
     transaction::{TransactionKind, TxEssenceEip1559, TxEssenceEip2930, TxEssenceLegacy},
-    transactions::{ethereum::EthereumTxEssence, Transaction, TxSignature},
+    transactions::{ethereum::EthereumTxEssence, Transaction},
     trie::{self, MptNode, MptNodeData, StateAccount},
     withdrawal::Withdrawal,
     Bloom, Bytes, RlpBytes, StorageKey, B160, B256, B64, U256, U64,
